Added Reactive Form for User Registration

+++ DatingApp.API/Controllers/PhotosController.cs
@@ -170,7 +170,7 @@ namespace DatingApp.API.Controllers
                 }
             }
 
-            
+            if(photoFromRepo.PublicId == null)
             {
                  _repo.Delete(photoFromRepo);
             }
